make purchases a LifecycleDelegate
