+++ purchases/src/main/kotlin/com/revenuecat/purchases/Purchases.kt
@@ -9,6 +9,7 @@ import android.annotation.SuppressLint
 import android.app.Activity
 import android.content.Context
 import androidx.annotation.VisibleForTesting
+import androidx.lifecycle.ProcessLifecycleOwner
 import com.revenuecat.purchases.common.LogIntent
 import com.revenuecat.purchases.common.PlatformInfo
 import com.revenuecat.purchases.common.debugLogsEnabled
@@ -45,7 +46,7 @@ typealias ErrorPurchaseCallback = (StoreTransaction, PurchasesError) -> Unit
  */
 class Purchases internal constructor(
     @get:JvmSynthetic internal val purchasesOrchestrator: PurchasesOrchestrator,
-) {
+) : LifecycleDelegate {
 
     internal var appConfig = purchasesOrchestrator.appConfig
 
@@ -88,6 +89,28 @@ class Purchases internal constructor(
      */
     val store: Store
         get() = purchasesOrchestrator.store
+
+    private val lifecycleHandler: AppLifecycleHandler by lazy {
+        AppLifecycleHandler(this)
+    }
+
+    init {
+        purchasesOrchestrator.dispatch {
+            // This needs to happen after the billing client listeners have been set. This is because
+            // we perform operations with the billing client in the lifecycle observer methods.
+            ProcessLifecycleOwner.get().lifecycle.addObserver(lifecycleHandler)
+        }
+    }
+
+    /** @suppress */
+    override fun onAppBackgrounded() {
+        purchasesOrchestrator.onAppBackgrounded()
+    }
+
+    /** @suppress */
+    override fun onAppForegrounded() {
+        purchasesOrchestrator.onAppForegrounded()
+    }
 
     // region Public Methods
 
@@ -384,6 +407,10 @@ class Purchases internal constructor(
      */
     fun close() {
         purchasesOrchestrator.close()
+
+        purchasesOrchestrator.dispatch {
+            ProcessLifecycleOwner.get().lifecycle.removeObserver(lifecycleHandler)
+        }
     }
 
     /**
